third_party/WebKit/LayoutTests/external/wpt/html/webappapis/scripting/events/event-handler-processing-algorithm-expected.txt

This is a testharness.js-based test.
FAIL mouseover listener returning true cancels event assert_equals: expected true but got false
FAIL mouseover listener returning false doesn't cancel event assert_equals: expected false but got true
PASS click listener returning false cancels event 
PASS blur listener returning false cancels event 
PASS dblclick listener returning false cancels event 
Harness: the test ran to completion.
